Replace psa-attestation/lib/mbed-crypto with mbedtls from rust-psa-crypto.
The crate rust-psa-crypto is slightly modified from upstream, primarily
to configure mbedtls differently: "baremetal" with entropy obtained with
the C-library function getrandom.

Also, rust-psa-crypto/Cargo.toml is removed to avoid nested workspaces.

This uses a different version of the C library with a slightly different
API. In particular, key handles are now uint32_t instead of uint16_t.
This accounts for most of the changes in psa-attestation and t_cose.

There are various IceCap changes to enable building rust-psa-crypto.
These include adding a bogus implementation of getrandom()!

+++ proxy-attestation-server/src/attestation/psa.rs
@@ -89,7 +89,7 @@ pub fn attestation_token(body_string: String) -> ProxyAttestationServerResponder
     let mut t_cose_ctx: t_cose_sign1_verify_ctx = unsafe { ::std::mem::zeroed() };
     unsafe { t_cose_sign1_verify_init(&mut t_cose_ctx, 0) };
 
-    let mut key_handle: u16 = 0;
+    let mut key_handle: u32 = 0;
     let lpk_ret = unsafe {
         t_cose_sign1_verify_load_public_key(
             &PUBLIC_KEY as *const u8,
